Daybreak 0.9.8.18 (#179)
* Minimap implementation

* Fix minimap

* Add debounce logic to entities positions

* Cleanup

* Fix x64 build memory reading
Move build to x64
Fix entity debouncing
Added some backstops to prevent reading corrupted memory

* Add minimap zooming and panning functionality
Add main window position saving procedure

* Increment version

---------

+++ Daybreak/Models/Interop/GuildwarsArray.cs
@@ -6,7 +6,7 @@ namespace Daybreak.Models.Interop;
 [StructLayout(LayoutKind.Sequential)]
 public readonly struct GuildwarsArray
 {
-    public readonly IntPtr Buffer;
+    public readonly int Buffer;
 
     public readonly uint Capacity;
